# perf(ui): improve merging of continuous name match ranges