### main
```
results_list            time:   [17.800 µs 17.836 µs 17.869 µs]
                        change: [-0.7235% -0.4176% -0.1185%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
```

### this branch
```
results_list            time:   [12.780 µs 12.789 µs 12.798 µs]
                        change: [-28.381% -28.218% -28.059%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe
```